ai-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
crewai==0.30.11
crewai-tools==0.2.6
openai==1.30.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
chromadb==0.5.0
sentence-transformers==2.7.0
requests==2.31.0
